% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_column.R
\name{check_column_name}
\alias{check_column_name}
\title{Upload Excel template}
\usage{
check_column_name(data, skip.col = 1)
}
\arguments{
\item{data}{List containing important tables.}

\item{skip.col}{= 1}
}
\description{
Function to upload an Excel file based on the predefined template
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=upload_gabarit_ADN]{upload_gabarit_ADN()}} to create the list of tables.

List references
}
