% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_column.R
\name{check_column_values_ID}
\alias{check_column_values_ID}
\title{Check column values}
\usage{
check_column_values_ID(data)
}
\arguments{
\item{data}{List containing important tables.}
}
\description{
Function to check that column name and order follow the predefined template
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=correct_column_name]{correct_column_name()}} to create the list of tables.

List references
}
