% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_key.R
\name{check_key}
\alias{check_key}
\title{Check that we are not trying to import data that should be already present}
\usage{
check_key(data, DB = "LabGeno")
}
\arguments{
\item{data}{List containing important tables.}
}
\description{
Function to check that column name and order follow the predefined template
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=upload_gabarit_ADN]{upload_gabarit_ADN()}} to create the list of tables.

List references
}
