% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_key.R
\name{check_key_unique}
\alias{check_key_unique}
\title{Internal function to correct key column by column}
\usage{
check_key_unique(data, DB, table, DB.table, key)
}
\arguments{
\item{data}{a string}

\item{DB}{a string}

\item{table}{a string}

\item{DB.table}{a string}

\item{key}{a string}
}
\description{
Function to correct some weird format of latitude and longitude
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
