% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_relations.R
\name{check_relation_ADNe}
\alias{check_relation_ADNe}
\title{Check relations between tables for the eDNA gabarit}
\usage{
check_relation_ADNe(data)
}
\arguments{
\item{data}{List containing important tables.}
}
\description{
Function to check that relationship between table are respected
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=upload_gabarit_ADN]{upload_gabarit_ADN()}} to create the list of tables.

List references
}
