% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_relations.R
\name{check_relation_wDB}
\alias{check_relation_wDB}
\title{Check relations between tables, including a connection with the DB}
\usage{
check_relation_wDB(data, DB = "LabGeno")
}
\arguments{
\item{data}{List containing important tables.}

\item{DB}{Name of the ODBC connection}
}
\description{
Function to check that column name and order follow the predefined template
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=upload_gabarit_ADN]{upload_gabarit_ADN()}} to create the list of tables.

List references
}
