% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upload_gabarit.R
\name{combine_multiple_gabarit}
\alias{combine_multiple_gabarit}
\title{Upload and combine multiple corrected tables}
\usage{
combine_multiple_gabarit(
  path,
  table.access = c("Groupes", "Specimens", "Tissus", "Extraits_ADN_ARN",
    "Analyse_Externe", "Sexage", "Sequencage", "Hormones")
)
}
\arguments{
\item{path}{Path to the folder were the excel files are}

\item{table.access}{List of sheets that will be read (only if they exist)}
}
\description{
Function to upload and combined corrected Excel files
}
\details{

}
\examples{
# provide some examples of how to use your function

}
\seealso{
\code{\link[=upload_gabarit_ADN]{upload_gabarit_ADN()}}.

List references
}
