% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upload_gabarit.R
\name{combine_multiple_gabarit_ADNe}
\alias{combine_multiple_gabarit_ADNe}
\title{Upload and combine multiple corrected tables}
\usage{
combine_multiple_gabarit_ADNe(
  path,
  table.access = c("Sites_ADNe", "Stations_ADNe", "Echantillons_ADNe", "Filtres_ADNe",
    "Extraits_ADNe", "qPCR_ADNe", "qPCR_inhibition_ADNe", "QNC_QPC_ADNe",
    "Librairies_ADNe", "Purification_librairies_ADNe", "Analyses_externes_librairies_AD",
    "Courbe_etalonnage_ADNe", "Sequencage_Sanger_ADNe")
)
}
\arguments{
\item{path}{Path to the folder were the excel files are}

\item{table.access}{List of sheets that will be read (only if they exist)}
}
\description{
Function to upload and combined corrected Excel files
}
\examples{
# provide some examples of how to use your function

}
\seealso{
\code{\link[=upload_gabarit_ADN]{upload_gabarit_ADN()}}.

List references
}
