% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_column.R
\name{correct_column_name}
\alias{correct_column_name}
\title{Check column names}
\usage{
correct_column_name(data)
}
\arguments{
\item{data}{List containing important tables.}
}
\description{
Function to check that column name and order follow the predefined template
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=upload_gabarit_ADN]{upload_gabarit_ADN()}} to create the list of tables.

List references
}
