% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_column.R
\name{correct_column_values_hour}
\alias{correct_column_values_hour}
\title{Check and correct column with hour}
\usage{
correct_column_values_hour(data)
}
\arguments{
\item{data}{List containing important tables.}
}
\description{
Function to check that column with hour are in the right format, and do a correction if needed.
}
\details{
Character : will be changed for h
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=correct_column_name]{correct_column_name()}} to create the list of tables.

List references
}
