% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{correct_latlon}
\alias{correct_latlon}
\title{Correct lat - long}
\usage{
correct_latlon(x, format = c("NGSL", "dms"))
}
\arguments{
\item{x}{a string}
}
\description{
Function to correct some weird format of latitude and longitude
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# In the Tel format
x <- 4753.1
correct_latlon(x, format = "NGSL")

# In the dms format
x <- "47° 22' 38,121 N"
correct_latlon(x, format = "dms")

}
