% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_primer.R
\name{correct_primer}
\alias{correct_primer}
\title{Check the Primer columns specifically}
\usage{
correct_primer(data, DB = "LabGeno")
}
\arguments{
\item{data}{List containing important tables.}

\item{DB}{Name of the DB in the ODBC.}
}
\description{
Function to check that primers and probes names are in the right format
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=correct_column_name]{correct_column_name()}} to create the list of tables.

List references
}
