% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_tables.R
\name{correct_soumissionGQ}
\alias{correct_soumissionGQ}
\title{Check the Numero de Soumissions GQ columns specifically}
\usage{
correct_soumissionGQ(data, DB = "LabGeno")
}
\arguments{
\item{data}{List containing important tables.}

\item{DB}{Name of the DB in the ODBC.}
}
\description{
Function to check that No_soumission_GQ are in the right format
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=correct_column_name]{correct_column_name()}} to create the list of tables.

List references
}
