% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_results.R
\name{export_access_xlsx}
\alias{export_access_xlsx}
\title{Upload Excel template}
\usage{
export_access_xlsx(data, path = ".", prefix = "Export")
}
\arguments{
\item{data}{Path to the excel spreadsheet}

\item{path}{Path to the excel spreadsheet}

\item{prefix}{Prefi of the new file name}
}
\description{
Function to export the list into a csv
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=upload_gabarit_ADN]{upload_gabarit_ADN()}}

List references
}
