% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_ADNe_V3.R
\name{extract_libraries_ADNe}
\alias{extract_libraries_ADNe}
\title{Extract old eDNA gabarit Purification_librairies_ADNe et Analyses_externes_librairies_ADNe info}
\usage{
extract_libraries_ADNe(data)
}
\arguments{
\item{data}{List containing important tables.}
}
\description{
Function to extract the Purification_librairies_ADNe and Analyses_externes_librairies_ADNe info from the qPCR table, as encounter in the eDNA gabarit v3.
}
\details{
The data object should include a "Librairie_ADNe" table, but not a "Purification_librairies_ADNe" and "Analyses_externes_librairies_ADNe"
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=upload_gabarit_ADNe]{upload_gabarit_ADNe()}} to create the list of tables.

List references
}
