% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_ADNe_V3.R
\name{extract_qpc_qnc_ADNe}
\alias{extract_qpc_qnc_ADNe}
\title{Extract old eDNA gabarit QNC QPC sequencing info}
\usage{
extract_qpc_qnc_ADNe(data)
}
\arguments{
\item{data}{List containing important tables.}
}
\description{
Function to extract the QPC QNC info from the qPCR table, as encounter in the eDNA gabarit v3.
}
\details{
The data object should include a "qPCR_ADNe" table, but not a "QNC_QPC_ADNe" table.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=upload_gabarit_ADNe]{upload_gabarit_ADNe()}} to create the list of tables.

List references
}
