% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect_DB.R
\name{list_DB}
\alias{list_DB}
\title{List table and request contained in the database.}
\usage{
list_DB(DB = "LabGeno")
}
\arguments{
\item{DB}{Name of the OBDC database, as set in your computer.}
}
\description{
Function to ;ist possible main table and request contained in the database.
}
\examples{
list_DB()

}
\seealso{
List references
}
