% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect_DB.R
\name{load_columns_DB}
\alias{load_columns_DB}
\title{Upload specific column of a DB table}
\usage{
load_columns_DB(columns, table, DB = "LabGeno", verbose = TRUE)
}
\arguments{
\item{columns}{Vector of the columne to upload.}

\item{table}{Name of the table you want to load.}

\item{DB}{Name of the OBDC database, as set in your computer.}

\item{verbose}{Print info if TRUE.}
}
\description{
Function to upload one or more specific columns from a the table in database
}
\details{
Possible tables are :  00_Liste_projets, 01_Groupes, 02_Specimens, 03_Tissus, 04_Extraits_ADN_ARN, 05_Analyse_Externe, 06_Sexage, 07_Sequencage, 08_Hormones, 09_GQ
}
\examples{
# Not to run
res <-load_columns_DB(columns = c("Numero_unique_groupe"), table = "01_Groupes")
head(res)
}
\seealso{
List references
}
