% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model.eDNA_convert}
\alias{model.eDNA_convert}
\title{Gabbarit to convert eDNA columns from V3 to V4 gabarit}
\format{
A tibble with 30 rows and 8 variables
}
\usage{
data(model.eDNA_convert)
}
\description{
test

\itemize{
\item
}
}
