% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model.other}
\alias{model.other}
\title{Gabbarit other special columns changes}
\format{
A tibble with 30 rows and 8 variables
}
\usage{
data(model.list)
}
\description{
test

\itemize{
\item
}
}
