% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{remove_all_special}
\alias{remove_all_special}
\title{Remove all special caracterd}
\usage{
remove_all_special(x)
}
\arguments{
\item{x}{a string}
}
\description{
Function to check that column name and order follow the predefined template
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function

}
