% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect_DB.R
\name{test_DB}
\alias{test_DB}
\title{Test connection with DB}
\usage{
test_DB(DB = "LabGeno")
}
\arguments{
\item{DB}{Name of the OBDC database, as set in your computer.}
}
\description{
Function to check that the connection with the DB is OK
}
\details{
The connection to your DB should have been set previously on your computer with ODBC manager.
}
\examples{
# Not to run
test_DB()

}
\seealso{
\code{\link[=load_DB]{load_DB()}} to upload a specific table

List references
}
