% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upload_gabarit.R
\name{upload_gabarit_ADN}
\alias{upload_gabarit_ADN}
\title{Upload Excel template}
\usage{
upload_gabarit_ADN(path, skip = 1)
}
\arguments{
\item{path}{Path to the excel spreadsheet}

\item{skip}{N row to skip}
}
\description{
Function to upload an Excel file based on the predefined template
}
\details{
Optional if you want to add more detais.
}
\examples{
# provide some examples of how to use your function
hello()

}
\seealso{
List relevant other functions \code{\link[=littleforecast]{littleforecast()}}.

List references
}
