% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upload_gabarit.R
\name{upload_gabarit_ADNe}
\alias{upload_gabarit_ADNe}
\title{Upload Excel eDNA template}
\usage{
upload_gabarit_ADNe(
  path,
  skip = 0,
  site = "Sites_ADNe",
  station = "Stations_ADNe",
  echantillon = "Echantillons_ADNe",
  filtre = "Filtres_ADNe",
  extrait = "Extraits_ADNe",
  qpcr_inhibition = "qPCR_inhibition_ADNe",
  qpcr = "qPCR_ADNe",
  qnc_qpc = "QNC_QPC_ADNe",
  sequencage = "Sequencage_Sanger_ADNe",
  courbe = "Courbe_etalonnage_ADNe",
  librairies = "Librairies_ADNe",
  purification = "Purification_librairies_ADNe",
  analyses_ext_lib = "Analyses_externes_librairies_ADNe",
  librairies_seq = "Librairies_SeqReady_ADNe"
)
}
\arguments{
\item{path}{Path to the excel spreadsheet}

\item{skip}{N row to skip. Default value is 0.}

\item{site}{Name of the sheet in the Excel file containing site data}

\item{station}{Name of the sheet in the Excel file containing station data}

\item{echantillon}{Name of the sheet in the Excel file containing sample data}

\item{filtre}{Name of the sheet in the Excel file containing the filter data}

\item{extrait}{Name of the sheet in the Excel file containing extract data}

\item{qpcr_inhibition}{of the sheet in the Excel file containing the qpcr inihibiton info}

\item{qpcr}{Name of the sheet in the Excel file containing the qpcr results}

\item{qnc_qpc}{Name of the sheet in the Excel file containing the qpcr control info}

\item{sequencage}{Name of the sheet in the Excel file containing the Sanger sequencing info}

\item{courbe}{Name of the sheet in the Excel file containing the standard curve info}

\item{librairies}{Name of the sheet in the Excel file containing the librairy info}

\item{purification}{Name of the sheet in the Excel file containing the librairy purification info}

\item{analyses_ext_lib}{Name of the sheet in the Excel file containing the librairy sequencing info}

\item{librairies_seq}{Name of the sheet in the Excel file containing the Librairie SeqReady ADNe info}
}
\description{
Function to upload an Excel file based on the predefined eDNA template
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
List references
}
