% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_projet.R
\name{correct_project}
\alias{correct_project}
\title{Check the Projet column specificallyr}
\usage{
correct_project(data, model.path = "inst/LISTE_PROJETS.xlsx")
}
\arguments{
\item{data}{List containing important tables.}

\item{path}{Path to an excel spreadsheet with the list of Projet name.
'}
}
\description{
Function to check that column name and order follow the predefined template
}
\details{
NULL if you doesn't want a specif sheet to be uploaded. No other check than sheet name.
}
\examples{
# provide some examples of how to use your function


}
\seealso{
\code{\link[=check_column_name]{check_column_name()}} to create the list of tables.

List references
}
